# backend/swagger_server/models/response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Response(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, type: str=None, message: str=None):  # noqa: E501
        """Response - a model defined in Swagger

        :param type: The type of this Response.  # noqa: E501
        :type type: str
        :param message: The message of this Response.  # noqa: E501
        :type message: str
        """
        self.swagger_types = {
            'type': str,
            'message': str
        }

        self.attribute_map = {
            'type': 'type',
            'message': 'message'
        }
        self._type = type
        self._message = message

    @classmethod
    def from_dict(cls, dikt) -> 'Response':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Response of this Response.  # noqa: E501
        :rtype: Response
        """
        return util.deserialize_model(dikt, cls)

    @property
    def type(self) -> str:
        """Gets the type of this Response.


        :return: The type of this Response.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this Response.


        :param type: The type of this Response.
        :type type: str
        """

        self._type = type

    @property
    def message(self) -> str:
        """Gets the message of this Response.


        :return: The message of this Response.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this Response.


        :param message: The message of this Response.
        :type message: str
        """

        self._message = message