# backend/swagger_server/models/tablet.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Tablet(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, input: str=None):  # noqa: E501
        """Tablet - a model defined in Swagger

        :param input: The input of this Tablet.  # noqa: E501
        :type input: str
        """
        self.swagger_types = {
            'input': str
        }

        self.attribute_map = {
            'input': 'input'
        }
        self._input = input

    @classmethod
    def from_dict(cls, dikt) -> 'Tablet':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Tablet of this Tablet.  # noqa: E501
        :rtype: Tablet
        """
        return util.deserialize_model(dikt, cls)

    @property
    def input(self) -> str:
        """Gets the input of this Tablet.


        :return: The input of this Tablet.
        :rtype: str
        """
        return self._input

    @input.setter
    def input(self, input: str):
        """Sets the input of this Tablet.


        :param input: The input of this Tablet.
        :type input: str
        """

        self._input = input