# backend/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: ai-challenge
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: http://localhost:8080
tags:
- name: Device
  externalDocs:
    description: Find out more
    url: http://swagger.io
paths:
  /laptop:
    post:
      tags:
      - Device
      summary: send laptop information
      operationId: laptop_post
      requestBody:
        description: laptop info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Laptop'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      x-openapi-router-controller: swagger_server.controllers.device_controller
  /smartphone:
    post:
      tags:
      - Device
      summary: send smartphone information
      operationId: smartphone_post
      requestBody:
        description: smartphone info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Smartphone'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      x-openapi-router-controller: swagger_server.controllers.device_controller
  /tablet:
    post:
      tags:
      - Device
      summary: send Tablet information
      operationId: tablet_post
      requestBody:
        description: smartphone info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tablet'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      x-openapi-router-controller: swagger_server.controllers.device_controller
  /smartwatch:
    post:
      tags:
      - Device
      summary: send smartwatch information
      operationId: smartwatch_post
      requestBody:
        description: smartwatch info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Smartwatch'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      x-openapi-router-controller: swagger_server.controllers.device_controller
  /Desktop:
    post:
      tags:
      - Device
      summary: send Desktop information
      operationId: desktop_post
      requestBody:
        description: Desktop info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Desktop'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      x-openapi-router-controller: swagger_server.controllers.device_controller
components:
  schemas:
    Laptop:
      type: object
      properties:
        input:
          type: string
      xml:
        name: Laptop
    Smartwatch:
      type: object
      properties:
        input:
          type: string
      xml:
        name: Smartwatch
    Desktop:
      type: object
      properties:
        input:
          type: string
      xml:
        name: Desktop
    Smartphone:
      type: object
      properties:
        input:
          type: string
      xml:
        name: Smartphone
    Tablet:
      type: object
      properties:
        input:
          type: string
      xml:
        name: Tablet
    Response:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
      example:
        type: type
        message: message
      xml:
        name: Response